+++ frontend/src/pages/MainPage.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { ROUTES } from '../constants/routes';
-import useAuthConditionTasks from '../hooks/@common/useAuthConditionTasks';
+import useAuthConditionTasks from '../hooks/domain/auth/useAuthConditionTasks';
 
 const MainPage = () => {
   const navigate = useNavigate();
