+++ frontend/src/pages/create/funnel/SpaceCreateFunnel.tsx
@@ -3,9 +3,9 @@ import { useNavigate } from 'react-router-dom';
 import { DiamondImg as diamondImage } from '../../../@assets/images';
 import StepProgressBar from '../../../components/@common/progressBar/step/StepProgressBar';
 import { ROUTES } from '../../../constants/routes';
-import useAuthConditionTasks from '../../../hooks/@common/useAuthConditionTasks';
 import useConfirmBeforeRefresh from '../../../hooks/@common/useConfirmBeforeRefresh';
-import useAgreements from '../../../hooks/domain/useAgreements';
+import useAgreements from '../../../hooks/domain/auth/useAgreements';
+import useAuthConditionTasks from '../../../hooks/domain/auth/useAuthConditionTasks';
 import useFunnelHistory from '../../../hooks/useFunnelHistory';
 import type { SpaceFunnelInfo } from '../../../types/space.type';
 import AccessTypeElement from '../funnelElements/accessTypeElement/AccessTypeElement';
