+++ frontend/src/pages/create/funnelElements/checkSpaceInfoElement/CheckSpaceInfoElement.tsx
@@ -4,8 +4,8 @@ import LeftTimeInformationBox from '../../../../components/specific/leftTimeInfo
 import { INFORMATION } from '../../../../constants/messages';
 import { ROUTES } from '../../../../constants/routes';
 import useLeftTimer from '../../../../hooks/@common/useLeftTimer';
-import useAgreements from '../../../../hooks/domain/useAgreements';
-import useCreateSpace from '../../../../hooks/useCreateSpace';
+import useAgreements from '../../../../hooks/domain/auth/useAgreements';
+import useCreateSpace from '../../../../hooks/domain/space/useCreateSpace';
 import type { FunnelElementProps } from '../../../../types/funnel.type';
 import type { SpaceFunnelInfo } from '../../../../types/space.type';
 import { calculateKstToday } from '../../../../utils/calculateKstToday';
