+++ frontend/src/pages/guest/imageUploadPage/ImageUploadPage.tsx
@@ -17,12 +17,12 @@ import UploadBox from '../../../components/specific/uploadBox/UploadBox';
 import { ROUTES } from '../../../constants/routes';
 import { useOverlay } from '../../../contexts/OverlayProvider';
 import useLeftTimer from '../../../hooks/@common/useLeftTimer';
-import useLocalFile from '../../../hooks/@common/useLocalFile';
+import useGuestNickName from '../../../hooks/domain/guest/useGuestNickName';
+import useFileUpload from '../../../hooks/domain/photos/useFileUpload';
+import useLocalFile from '../../../hooks/domain/photos/useLocalFile';
+import useSpaceCodeFromPath from '../../../hooks/domain/space/useSpaceCodeFromPath';
+import useSpaceInfo from '../../../hooks/domain/space/useSpaceInfo';
 import useScrollUITriggers from '../../../hooks/domain/ui/useScrollUITriggers';
-import useFileUpload from '../../../hooks/useFileUpload';
-import useGuestNickName from '../../../hooks/useGuestNickName';
-import useSpaceCodeFromPath from '../../../hooks/useSpaceCodeFromPath';
-import useSpaceInfo from '../../../hooks/useSpaceInfo';
 import { ScrollableBlurArea } from '../../../styles/@common/ScrollableBlurArea.styles';
 import { theme } from '../../../styles/theme';
 import { checkIsEarlyDate } from '../../../utils/checkIsEarlyTime';
