+++ frontend/src/pages/landing/LandingPage.tsx
@@ -20,10 +20,10 @@ import Footer from '../../components/layout/global/footer/Footer';
 import KakaoLoginButton from '../../components/specific/kakaoLoginButton/KakaoLoginButton';
 import LeftTimeInformationBox from '../../components/specific/leftTimeInformationBox/LeftTimeInformationBox';
 import { AUTH_COOKIES } from '../../constants/keys';
-import useAuthActions from '../../hooks/@common/useAuthActions';
 import useLandingScroll from '../../hooks/@common/useLandingScroll';
 import useLeftTimer from '../../hooks/@common/useLeftTimer';
-import useKakaoAuth from '../../hooks/domain/useKakaoAuth';
+import useAuthActions from '../../hooks/domain/auth/useAuthActions';
+import useKakaoAuth from '../../hooks/domain/auth/useKakaoAuth';
 import { theme } from '../../styles/theme';
 import { CookieUtils } from '../../utils/CookieUtils';
 import { formatDate } from '../../utils/formatDate';
